Rename file before upload
The purpose of this commit is to give each stream a unique filename
since the token is per user and thus will be reused.  That means if
a user were to stream again the original file would be clobbered on S3.
I considered using a versioned bucket to get around that, but adding
timestamps to the filename is also useful metadata which is more
accurate than trying to infer the same information from the s3
timestamps in case there is any lag in the system.